# api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Internal HTTP API",
    "version": "1.0.0-alpha",
    "summary": "Public reader endpoints, editorial writes behind a bearer JWT, and the operational probes served by the Go binary.",
...
  },
  "servers": [
    {
      "url": "http://api:8080",
      "description": "The internal service address inside compose and Kubernetes; the Astro server reaches it as API_BASE_URL."
    },
    {
      "url": "http://localhost:8080",
      "description": "A locally run binary (HTTP_ADDR defaults to :8080)."
    }
  ],
  "security": [],
...
  "paths": {
...
    "/api/v1/cashback/withdrawals": {
      "post": {
        "tags": [
          "cashback-member"
        ],
        "operationId": "requestCashbackWithdrawal",
        "summary": "Ask to be paid",
        "description": "Reserve confirmed cashback and record a request for an operator to approve (US4, FR-050, FR-051).\n\nThe reservation happens in the SAME transaction as the request (D9), before any human reviews it, because the double-spend window is between request and approval. Two concurrent requests for the whole balance therefore produce one 201 and one 409: the second reads what the first left.\n\nThe member is the bearer of the token. There is no path segment and no query parameter naming an account, so asking to be paid out of somebody else's balance is not a request this endpoint can express.\n\nThe whole path is unmounted when cashback is off, and then answers 404.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashbackWithdrawalRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The request, with what was reserved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashbackWithdrawal"
                }
              }
            }
          },
          "400": {
            "description": "The body is not a single JSON document of the expected shape, the amount is not a positive {minor, currency}, or the currency is not one this deployment pays out in.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "description": "The destination is not the caller's. A destination that does not exist answers the same way, so an id cannot be probed for existence one request at a time.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request cannot be met. `code` is `insufficient_confirmed_balance` - carrying `shortfall`, and `threshold`/`confirmed` or `confirmed`/`requested` - or `destination_not_verified` for a destination of the caller's own that has not been proved theirs (FR-051).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InsufficientConfirmedBalance"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "description": "This deployment has not configured a withdrawal threshold or the house account earnings are paid out of. Nothing is broken; the deployment is incomplete, and the log names the keys.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cashback-member"
        ],
        "operationId": "listWithdrawals",
        "summary": "List the caller's withdrawal requests",
        "description": "Every withdrawal this member has asked for, newest first, with the operator's decision where one has been made and the rail's reference where a payment has been submitted.\n\nThe member is the bearer of the token; reading somebody else's is not a request this endpoint can express.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The member's withdrawal requests.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberWithdrawalPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/cashback/withdrawals/{id}": {
      "get": {
        "tags": [
          "cashback-member"
        ],
        "operationId": "getWithdrawal",
        "summary": "Read one withdrawal request",
        "description": "One request of the caller's own.\n\nA request belonging to somebody else and one that does not exist both answer 404, so an id cannot be probed for existence one request at a time.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberWithdrawal"
                }
              }
            }
          },
          "400": {
            "description": "The id in the path is not a UUID.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "404": {
            "description": "No withdrawal request of the caller's has that id.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/cashback/payout-destinations": {
      "get": {
        "tags": [
          "cashback-member"
        ],
        "operationId": "listPayoutDestinations",
        "summary": "List the caller's payout destinations",
        "description": "Every destination this member has recorded, oldest first, with whether each has been proved theirs (FR-051).\n\nNo bank details appear in the response and none ever will: they live outside this service (ADR-0003).\n\nThe member is the bearer of the token; there is no path segment or query parameter naming an account.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The member's destinations.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutDestinationPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": [
          "cashback-member"
        ],
        "operationId": "recordPayoutDestination",
        "summary": "Record somewhere to be paid",
        "description": "Record a payout destination. It arrives UNVERIFIED, always: verification is a separate flow (FR-051), and one that arrived verified would be one nobody proved belongs to the member.\n\nThe details are handed straight to this deployment's details store and never held here; what is recorded is a reference (ADR-0003). They are not echoed in the response, not logged, and not quoted in a refusal.\n\nThe member is the bearer of the token; there is no field naming an account.\n\n503 when the deployment has no details store configured. Only this endpoint needs one: listing destinations and withdrawing to those that already exist keep working.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPayoutDestination"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The destination, unverified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutDestination"
                }
              }
            }
          },
          "400": {
            "description": "The body is not a single JSON document of the expected shape, the kind is not one this deployment pays out on, the details are missing, or the details store refused them. The refusal never quotes what was sent.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "The details store did not answer. Nothing was recorded, so the request can be repeated.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "This deployment has no details store configured, so it has nowhere to put bank details. Nothing is broken; the deployment is incomplete, and the log says so.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A JWT from the auth provider, verified against the configured JWKS. The subject must map to an account. Endpoints under /api/v1/editorial/ additionally require that account to hold the editor role, and the database enforces that rule again on write; endpoints under /api/v1/account/ require only that somebody is signed in, because they act on the caller's own row."
      }
    },
...
    "schemas": {
...
      "CashbackWithdrawalRequest": {
        "type": "object",
        "description": "A member asking to be paid their cashback. The member is the bearer of the token and is deliberately absent from the body: there is no field naming an account, so asking to be paid out of somebody else's balance is not a request this endpoint can express.",
        "required": [
          "destination_id",
          "amount"
        ],
        "additionalProperties": false,
        "properties": {
          "destination_id": {
            "type": "string",
            "format": "uuid",
            "description": "One of the caller's own verified payout destinations (FR-051). A destination belonging to somebody else answers 403, and so does one that does not exist."
          },
          "amount": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Money"
              }
            ],
            "description": "What the member is asking for. What is actually reserved may be more - see reserved_amount."
          }
        }
      },
      "CashbackWithdrawal": {
        "type": "object",
        "description": "A cashback withdrawal request as it was recorded, with the reservation already made (D9).",
        "required": [
          "request_id",
          "state",
          "reserved_amount"
        ],
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "awaiting_approval"
            ],
            "description": "Always awaiting_approval on creation: the money is reserved before any human reviews the request, and the review is an operator's act (C-4)."
          },
          "reserved_amount": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Money"
              }
            ],
            "description": "What was actually moved out of confirmed, which is at least the amount asked for. Cashback is reserved in whole entries - each cites the network report evidencing it (C-2), so there is no half of one to take - and a member who asks for slightly less than an entry has slightly more reserved. This, not the amount requested, is what a payout will pay."
          }
        }
      },
...
    }
  }
}